making tdd more organized and added mocked data for  user_event_enroll
