+++ backend/create_db.py
@@ -1,5 +1,5 @@
 from backend import app, db
-from mock_data import event_info_mocking, filter_info_mocking, club_info_mocking, user_info_mocking, review_rating_mocking
+from mock_data import event_info_mocking, filter_info_mocking, club_info_mocking, user_info_mocking, review_rating_mocking, user_event_enroll
 
 """
 DO NOT REMOVE THE FOLLOWING LINES, AND PLEASE EXECUTE WITH CAUTION
@@ -25,5 +25,6 @@ with app.app_context():
     club_info_mocking(db)
     user_info_mocking(db)
     review_rating_mocking(db)
+    user_event_enroll(db)
