+++ backend/models/user_enroll_event_mock_data.py
@@ -1,0 +1,12 @@
+
+user_enroll_event_mock_data = [
+    {
+        "event_id": 1,
+        "user_id": 1
+    },
+    {
+        "event_id": 2,
+        "user_id": 1
+    }
+]
+
